feat: update reset password to tsx
